+++ test/resources/pipe-modules/filter-t-async.js
@@ -1,0 +1,13 @@
+module.exports.filterT = (data) => {
+  return new Promise(resolve => {
+    let filteredData = {data: []};
+
+    for (const o of data.data) {
+      if (o.id.match(/T/)) {
+        filteredData.data.push(o);
+      }
+    }
+
+    resolve(filteredData);
+  });
+};